Wired in function call syntax
This required peeking ahead in the lexer to see if we had a left paren
directly after an identifier (ignoring whitespace). This creates a
subtle incompatibility with gawk in which builtin functions can be
called as in C with a space after the function name (e.g.
split (x, y, z)), but UDFs cannot. For now I think this is fine but
let's watch out for it.

This led to some basic bug fixes around builtin function calls. There
are outstanding bugs around OOB registers being generated for maps.