Fixed bug in side-effects.
Also added support for functions taking maps as arguments to propagate
constraints back to their args. This is needed for split, and it will be
important going forward if we add extra builtins. It is also good
practice for UDFs.

This change also exposes types to the tests in harness, allowing us to
assert that a certain minimal type was inferred, rather than just
String.