All tests so far now pass.
The most interesting bug fix in this latest round has been in how
dependencies are managed for UDFs. The usual mechanism of keeping an
extra hashmap from operands to the results of calls is not enough,
because the return type of a given function can change over the course
of constraint solving even if the arguments do not change.

To get around this, we now allocate a separate node for each (cfg-level)
UDF. This node has a new rule that "always fires," i.e. whenver it is
added to the worklist it will trigger all of its dependencies as well.
So now, for each UDF we have:

x = f(a, b);

create a node

base_f --- CallUDF(a, b, f) ---> x
  ^            |||
  |__________  return_f

Where CallUDF stands in for a function body, whose return node points
back to `base_f`. Thus, any change to any of f's return values will
trigger base_f, which will trigger x, which will in turn trigger a
reevaluation of the CallUDF constraint, propagating the new return
value.