Added new string impl.
This implementation uses less space than the existing one. It does this
without increasing the amount of indirection, or compromising on the
existing optimizations for strings.

The upside of this representation in addition to less memory and cheaper
moves is that it can be represented as a 128-bit integer scalar in LLVM,
which simplify some things. The downside is that the implementation is
less modular, and makes pervasive use of unsafe.