Completed initial support for multiple files
This involved a lot of plumbing and refactoring in the surrounding code.
The result is a bit slower for small-loop programs, but I think the
increased generality is enough of a boost that we can leave optimizing
it for a follow-up.

This just allows for reading the files; next/nextfile support will be
implemented in a follow-up change.