LLVM fast-path for string drops.
One of the most common operations called in frawk is the function to
drop a string. frawk strings are a discriminated union, with some
variants not requiring any explicit dropping at all, while others do.
This change implements the check for whether this explicit work for
drop (e.g. via deallocation) may be required as an LLVM function that
can then be inlined into the body of the frawk program.

The slow path still requires a function call, and there is probably more
that can be moved to LLVM in the future, but for now I've already seen
measurable improvements on workloads that create lots of small strings.