Implement client-side batching.
Did this after conducting some simple benchmarks showing that sending on
the channel had a high overhead in IO-heavy workloads compared with the
previous single-threaded BufWriter-based implementation. This new design
adds buffering back to the local FileHandles. The same benchmark shows
that CPU time is roughly equal for an IO-heavy benchmark, while wall
time is down by roughly 30% because the writes are now issued by a
separate thread.