Refactored UniqueStr and Instr
We need to make Instr Send, and in order to do that we needed to make
UniqueStr support Literal variants and not hard-code static lifetime.
We will probably need to use crossbeam's scope library to allow for
instrs to be sent accross threads, but it should at least be possible to
do so safely now.