Support trailing newlines in a program
This is the source for why many programs listed in #24 do not parse.
There is probably a more principled solution here that modifies the
grammar in LALRPOP, but most of the ones I tried led down dead-ends.
For now, we strip trailing newlines from the program at the lexing
stage.