Implement extra regex optimizations
This one rewrites patterns like  "^<literal>" to a function that does a
direct byte comparison. In the future we could do the same for "end of
line" anchors, or lower this further into a memcmp of some kind.

The performance gains from this change are modest. Much of the change
is spent implementing a "Handles" mechanism for ensuring certain
pointers will live as long as generated code (something that only
happens by accident today).