cleanup, bug fixes in type inference
I came accross this bug trying to diagnose a totally different bug where
we see a memory leak in the stat_2 benchmark. Still debugging that one.
I do not believe the changes in these branch caused that bug but I do
think that they make it more likely to trigger in normal programs.