Abandoning this change for now
The idea behind the changes in this branch was to detect statically when
certain fields in the input were only ever parsed as floating-point
values, and in those cases avoid "slicing to string". This doesn't
appear to help much; indeed, in the (partial, incomplete) version in
this branch, the reference benchmarks are slower.

That's all fine; and some of the techniques here may be worth returning
to in the future. But for now, I'm going to abandon development.