Add toupper and tolower builtin functions (#52)
* Add implementations for toupper and tolower

* Optimize map increments

* Fix bug in type inference around IncMap

It would be nice if we could have implemented this as a rewrite on the
typed AST. That would allow us to keep the existing structure around
type inference. However, that's a fairly cumbersome optimization to
implement in the current framework. I think it's worth exploring in the
future, though.

* Add toupper and tolower to the reference