Insert missing values on lookup
We've always had lookups return the default value for a given type when
the key is not present in the map, but we haven't actually been
inserting that key when performing the lookup. This change fixes that,
bringing frawk closer in line with Awk.

This change also does some minor refactoring of some of the handling of
shared mutability for maps.