Improve parsing of awk programs which have omitted curly braces around single statements
Improve parsing of awk programs which have single statements after a
`if`, `else`, `while` or `for` block but with omitted curly braces.

Those programs could not be parsed by frawk before and would give:

    Unrecognized token `}` found at line X, column Y:line X, column Y+1
    Expected one of "\n" or ";"

In awk statements are terminated by semicolons, newlines or right braces.
The current lalrpop grammar does not support parsing statements that are
terminated by a right brace from an upper level, so a ';' gets injected
before a '}' if none of the following conditions occurs:
  - Previous token is not a newline or ';'.
  - Previous token is not a '{', as "{}" should not be converted to invalid "{;}".